+++ packages/graph-explorer/src/core/ConnectedProvider/ConnectedProvider.tsx
@@ -3,7 +3,6 @@ import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { NotificationProvider } from "@/components/NotificationProvider";
 import Toast from "@/components/Toast";
 import AppStatusLoader from "@/core/AppStatusLoader";
-import type { RawConfiguration } from "@/core/ConfigurationProvider";
 import StateProvider from "@/core/StateProvider/StateProvider";
 import ThemeProvider from "@/core/ThemeProvider/ThemeProvider";
 import { MantineProvider } from "@mantine/core";
@@ -12,10 +11,6 @@ import { ExpandNodeProvider } from "@/hooks/useExpandNode";
 import { ErrorBoundary } from "react-error-boundary";
 import AppErrorPage from "@/core/AppErrorPage";
 import { TooltipProvider } from "@/components";
-
-export type ConnectedProviderProps = {
-  config?: RawConfiguration;
-};
 
 function exponentialBackoff(attempt: number): number {
   return Math.min(attempt > 1 ? 2 ** attempt * 1000 : 1000, 30 * 1000);
@@ -32,10 +27,7 @@ const queryClient = new QueryClient({
   },
 });
 
-const ConnectedProvider = (
-  props: PropsWithChildren<ConnectedProviderProps>
-) => {
-  const { config, children } = props;
+export default function ConnectedProvider({ children }: PropsWithChildren) {
   return (
     <ErrorBoundary FallbackComponent={AppErrorPage}>
       <QueryClientProvider client={queryClient}>
@@ -45,7 +37,7 @@ const ConnectedProvider = (
               <ThemeProvider>
                 <NotificationProvider component={Toast}>
                   <StateProvider>
-                    <AppStatusLoader config={config}>
+                    <AppStatusLoader>
                       <ExpandNodeProvider>{children}</ExpandNodeProvider>
                     </AppStatusLoader>
                   </StateProvider>
@@ -57,6 +49,4 @@ const ConnectedProvider = (
       </QueryClientProvider>
     </ErrorBoundary>
   );
-};
-
-export default ConnectedProvider;
+}